+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 -   **terra-file-browser** use components from `@angular/material` instead of our own deprecated ones.
 -   **terra-image-preview** use components from `@angular/material` instead of our own deprecated ones.
 -   **terra-node-tree** add support for mat-icons.
+-   **tcTooltip** added `TerraTooltipModule` to be able to individually include tooltip functionality.
 
 # 12.2.0 (01.11.2021)